fix: current user data subscription
