+++ frontend/src/hooks/subscriptions/useUserByTaskSubscription.ts
@@ -1,8 +1,8 @@
 import { GetCurrentUserQuery } from 'pages/projectDetail/getUser.gen'
-
 import { useUpdateUserByTaskSubScSubscription } from 'pages/projectDetail/projectDetail.gen'
 import { useAuthContext } from 'providers/AuthProvider'
 import { useEffect, useState } from 'react'
+import logger from 'utils/debugger/logger'
 
 type UseUpdateUserByTaskSubscription = {
   updateUserByTask: GetCurrentUserQuery['user'] | undefined
@@ -21,6 +21,7 @@ export const useUpdateUserByTaskSubscription = (): UseUpdateUserByTaskSubscripti
   useEffect(() => {
     if (!data) return
 
+    logger.debug(data.updateUserByTask)
     setUpdateUserByTask({ ...data.updateUserByTask })
   }, [data])
