+++ frontend/src/hooks/useGetCurrentUserData.ts
@@ -1,26 +1,15 @@
-import { useEffect, useState } from 'react'
-import {
-  GetCurrentUserQuery,
-  useGetCurrentUserLazyQuery,
-  useNewNotificationsByCreateProjectSubScSubscription,
-  useNewNotificationsSubScSubscription,
-  useUpdateUserDataByCertificationSubScSubscription,
-  useUpdateUserDataByInterestSubScSubscription,
-  useUpdateUserDataSubScSubscription,
-  useUpdateUserDataByCreateProjectSubscription,
-} from 'pages/projectDetail/getUser.gen'
+import { Dispatch, SetStateAction, useEffect, useState } from 'react'
+import { GetCurrentUserQuery, useGetCurrentUserLazyQuery } from 'pages/projectDetail/getUser.gen'
 import { useAuthContext } from 'providers/AuthProvider'
 import { Notifications } from 'types/notification'
-import logger from 'utils/debugger/logger'
-import { useUpdateUserByTaskSubscription } from './subscriptions/useUserByTaskSubscription'
-import toast from 'utils/toast/toast'
-import Exp from 'utils/exp/exp'
 
 type UseGetCurrentUserDataReturn = {
   getCurrentUser: ReturnType<typeof useGetCurrentUserLazyQuery>[0]
   currentUserData: GetCurrentUserQuery['user'] | undefined
   firebaseCurrentUser: ReturnType<typeof useAuthContext>['currentUser']
   notifications: Notifications
+  setUserData: Dispatch<SetStateAction<GetCurrentUserQuery['user'] | undefined>>
+  setNotifications: Dispatch<SetStateAction<Notifications>>
 }
 
 /**
@@ -28,7 +17,6 @@ type UseGetCurrentUserDataReturn = {
  */
 export const useGetCurrentUserData = (): UseGetCurrentUserDataReturn => {
   const { currentUser: firebaseCurrentUser } = useAuthContext()
-  const { updateUserByTask } = useUpdateUserByTaskSubscription()
   const [getCurrentUser, currentUserData] = useGetCurrentUserLazyQuery({
     onCompleted(data) {
       const notifications: Notifications = data.user.invitations.map(invitation => {
@@ -42,36 +30,6 @@ export const useGetCurrentUserData = (): UseGetCurrentUserDataReturn => {
   })
   const [userData, setUserData] = useState<GetCurrentUserQuery['user']>()
   const [notifications, setNotifications] = useState<Notifications>([])
-  const newNotifications = useNewNotificationsSubScSubscription({
-    variables: {
-      user_id: String(firebaseCurrentUser?.uid),
-    },
-  })
-  const newNotificationsByCreateProject = useNewNotificationsByCreateProjectSubScSubscription({
-    variables: {
-      user_id: String(firebaseCurrentUser?.uid),
-    },
-  })
-  const updateUserData = useUpdateUserDataSubScSubscription({
-    variables: {
-      user_id: String(firebaseCurrentUser?.uid),
-    },
-  })
-  const updateUserDataByCertification = useUpdateUserDataByCertificationSubScSubscription({
-    variables: {
-      user_id: String(firebaseCurrentUser?.uid),
-    },
-  })
-  const updateUserDataByInterest = useUpdateUserDataByInterestSubScSubscription({
-    variables: {
-      user_id: String(firebaseCurrentUser?.uid),
-    },
-  })
-  const updateUserDataByCreateProject = useUpdateUserDataByCreateProjectSubscription({
-    variables: {
-      user_id: String(firebaseCurrentUser?.uid),
-    },
-  })
 
   useEffect(() => {
     if (!firebaseCurrentUser) return
@@ -86,61 +44,12 @@ export const useGetCurrentUserData = (): UseGetCurrentUserDataReturn => {
     setUserData(currentUserData.data?.user)
   }, [currentUserData])
 
-  useEffect(() => {
-    if (!newNotifications.data) return
-    logger.debug(newNotifications.data)
-
-    const notifications: Notifications = newNotifications.data.newInvitation.map(invitation => {
-      return {
-        ...invitation.project,
-        createAt: invitation.created_at,
-      }
-    })
-    setNotifications(notifications)
-  }, [newNotifications.data])
-
-  useEffect(() => {
-    if (!newNotificationsByCreateProject.data) return
-    logger.debug(newNotificationsByCreateProject.data)
-
-    const notifications: Notifications =
-      newNotificationsByCreateProject.data.newInvitationByCreateProject.map(invitation => {
-        return {
-          ...invitation.project,
-          createAt: invitation.created_at,
-        }
-      })
-    setNotifications(notifications)
-  }, [newNotificationsByCreateProject.data])
-
-  useEffect(() => {
-    logger.debug(updateUserByTask)
-    if (!updateUserByTask) return
-
-    if (userData && Exp.toLevel(updateUserByTask.exp) > Exp.toLevel(userData.exp)) {
-      toast.success(
-        'レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！レベルアップ！！！！！',
-      )
-    }
-
-    setUserData(updateUserByTask)
-  }, [updateUserByTask])
-
-  useEffect(() => {
-    setUserData(updateUserData.data?.updateUserData)
-  }, [updateUserData.data])
-
-  useEffect(() => {
-    setUserData(updateUserDataByCertification.data?.updateUserDataByCertification)
-  }, [updateUserDataByCertification.data])
-
-  useEffect(() => {
-    setUserData(updateUserDataByInterest.data?.updateUserDataByInterest)
-  }, [updateUserDataByInterest.data])
-
-  useEffect(() => {
-    setUserData(updateUserDataByCreateProject.data?.projectCreate)
-  }, [updateUserDataByCreateProject.data])
-
-  return { getCurrentUser, currentUserData: userData, firebaseCurrentUser, notifications }
+  return {
+    getCurrentUser,
+    currentUserData: userData,
+    firebaseCurrentUser,
+    notifications,
+    setUserData,
+    setNotifications,
+  }
 }
